Remove JS-applied help block CSS
This causes layout issues in most of our use cases (e.g., with
`form-horizontal`) and is difficult to override in the CSS because it's
applied to the element directly.

I think it should just be applied manually in the target application's
CSS in any special cases where it's actually needed.