Allow for modification of passed control_col option
In the markdown_textarea cell, when we attempt to build the form group
using the provided rails-bootstrap-forms form object, that gem then
tries to call `#concat` on our passed in `control_col` option:

https://github.com/bootstrap-ruby/rails-bootstrap-forms/blob/v2.3.0/lib/bootstrap_form/form_builder.rb#L190

Unfortunately this now fails with the recently added frozen string
pragma. To prevent the error, we can supply a non-frozen string instead.
Hopefully we can get rid of the mutation upstream in the forms gem and
this just becomes a temporary solution.